♻️ text event refactor.
